Move nom into nom_core workspace
This allows for the re-export of not only nom-core, thus making
this a no-change to the api, it also allows for reexport of other community extensions, such as num-supreme
