+++ nom_core/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "nom_core"
+version = "0.1.0"
+edition = "2021"
+
+
+[features]
+alloc = []
+std = ["alloc", "memchr/std"]
+default = ["std"]
+docsrs = []
+
+[dependencies.memchr]
+version = "2.3"
+default-features = false
+[dev-dependencies]
+doc-comment = "0.3"
+proptest = "=1.0.0"
