+++ nom_core/src/bits/complete.rs
@@ -10,6 +10,7 @@ use crate::traits::{Input, ToUsize};
 ///
 /// # Example
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::bits::complete::take;
 /// # use nom::IResult;
 /// # use nom::error::{Error, ErrorKind};
@@ -107,6 +108,7 @@ where
 ///
 /// # Example
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::bits::complete::bool;
 /// # use nom::IResult;
 /// # use nom::error::{Error, ErrorKind};
