+++ nom_core/src/bytes/complete.rs
@@ -1,23 +1,26 @@
-//! Parsers recognizing bytes streams, streaming version
+//! Parsers recognizing bytes streams, complete input version
 
 use core::marker::PhantomData;
 
 use crate::error::ParseError;
 use crate::internal::{IResult, Parser};
 use crate::traits::{Compare, FindSubstring, FindToken, ToUsize};
+use crate::Complete;
 use crate::Emit;
 use crate::Input;
 use crate::OutputM;
-use crate::Streaming;
 
-/// Recognizes a pattern.
+/// Recognizes a pattern
 ///
 /// The input data will be compared to the tag combinator's argument and will return the part of
-/// the input that matches the argument.
+/// the input that matches the argument
+///
+/// It will return `Err(Err::Error((_, ErrorKind::Tag)))` if the input doesn't match the pattern
 /// # Example
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::{Error, ErrorKind}, Needed, IResult};
-/// use nom::bytes::streaming::tag;
+/// use nom::bytes::complete::tag;
 ///
 /// fn parser(s: &str) -> IResult<&str, &str> {
 ///   tag("Hello")(s)
@@ -25,8 +28,7 @@ use crate::Streaming;
 ///
 /// assert_eq!(parser("Hello, World!"), Ok((", World!", "Hello")));
 /// assert_eq!(parser("Something"), Err(Err::Error(Error::new("Something", ErrorKind::Tag))));
-/// assert_eq!(parser("S"), Err(Err::Error(Error::new("S", ErrorKind::Tag))));
-/// assert_eq!(parser("H"), Err(Err::Incomplete(Needed::new(4))));
+/// assert_eq!(parser(""), Err(Err::Error(Error::new("", ErrorKind::Tag))));
 /// ```
 pub fn tag<T, I, Error: ParseError<I>>(tag: T) -> impl Fn(I) -> IResult<I, I, Error>
 where
@@ -39,7 +41,7 @@ where
       e: PhantomData,
     };
 
-    parser.process::<OutputM<Emit, Emit, Streaming>>(i)
+    parser.process::<OutputM<Emit, Emit, Complete>>(i)
   }
 }
 
@@ -47,10 +49,13 @@ where
 ///
 /// The input data will be compared to the tag combinator's argument and will return the part of
 /// the input that matches the argument with no regard to case.
+///
+/// It will return `Err(Err::Error((_, ErrorKind::Tag)))` if the input doesn't match the pattern.
 /// # Example
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::{Error, ErrorKind}, Needed, IResult};
-/// use nom::bytes::streaming::tag_no_case;
+/// use nom::bytes::complete::tag_no_case;
 ///
 /// fn parser(s: &str) -> IResult<&str, &str> {
 ///   tag_no_case("hello")(s)
@@ -60,7 +65,7 @@ where
 /// assert_eq!(parser("hello, World!"), Ok((", World!", "hello")));
 /// assert_eq!(parser("HeLlO, World!"), Ok((", World!", "HeLlO")));
 /// assert_eq!(parser("Something"), Err(Err::Error(Error::new("Something", ErrorKind::Tag))));
-/// assert_eq!(parser(""), Err(Err::Incomplete(Needed::new(5))));
+/// assert_eq!(parser(""), Err(Err::Error(Error::new("", ErrorKind::Tag))));
 /// ```
 pub fn tag_no_case<T, I, Error: ParseError<I>>(tag: T) -> impl Fn(I) -> IResult<I, I, Error>
 where
@@ -73,7 +78,7 @@ where
       e: PhantomData,
     };
 
-    parser.process::<OutputM<Emit, Emit, Streaming>>(i)
+    parser.process::<OutputM<Emit, Emit, Complete>>(i)
   }
 }
 
@@ -83,11 +88,12 @@ where
 ///
 /// It doesn't consume the matched character.
 ///
-/// It will return a `Err::Incomplete(Needed::new(1))` if the pattern wasn't met.
+/// It will return a `Err::Error(("", ErrorKind::IsNot))` if the pattern wasn't met.
 /// # Example
 /// ```rust
-/// # use nom::{Err, error::ErrorKind, Needed, IResult};
-/// use nom::bytes::streaming::is_not;
+/// # use nom_core as nom;
+/// # use nom::{Err, error::{Error, ErrorKind}, Needed, IResult};
+/// use nom::bytes::complete::is_not;
 ///
 /// fn not_space(s: &str) -> IResult<&str, &str> {
 ///   is_not(" \t\r\n")(s)
@@ -95,8 +101,8 @@ where
 ///
 /// assert_eq!(not_space("Hello, World!"), Ok((" World!", "Hello,")));
 /// assert_eq!(not_space("Sometimes\t"), Ok(("\t", "Sometimes")));
-/// assert_eq!(not_space("Nospace"), Err(Err::Incomplete(Needed::new(1))));
-/// assert_eq!(not_space(""), Err(Err::Incomplete(Needed::new(1))));
+/// assert_eq!(not_space("Nospace"), Ok(("", "Nospace")));
+/// assert_eq!(not_space(""), Err(Err::Error(Error::new("", ErrorKind::IsNot))));
 /// ```
 pub fn is_not<T, I, Error: ParseError<I>>(arr: T) -> impl FnMut(I) -> IResult<I, I, Error>
 where
@@ -105,7 +111,7 @@ where
 {
   let mut parser = super::is_not(arr);
 
-  move |i: I| parser.process::<OutputM<Emit, Emit, Streaming>>(i)
+  move |i: I| parser.process::<OutputM<Emit, Emit, Complete>>(i)
 }
 
 /// Returns the longest slice of the matches the pattern.
@@ -113,13 +119,12 @@ where
 /// The parser will return the longest slice consisting of the characters in provided in the
 /// combinator's argument.
 ///
-/// # Streaming specific
-/// *Streaming version* will return a `Err::Incomplete(Needed::new(1))` if the pattern wasn't met
-/// or if the pattern reaches the end of the input.
+/// It will return a `Err(Err::Error((_, ErrorKind::IsA)))` if the pattern wasn't met.
 /// # Example
 /// ```rust
-/// # use nom::{Err, error::ErrorKind, Needed, IResult};
-/// use nom::bytes::streaming::is_a;
+/// # use nom_core as nom;
+/// # use nom::{Err, error::{Error, ErrorKind}, Needed, IResult};
+/// use nom::bytes::complete::is_a;
 ///
 /// fn hex(s: &str) -> IResult<&str, &str> {
 ///   is_a("1234567890ABCDEF")(s)
@@ -128,8 +133,8 @@ where
 /// assert_eq!(hex("123 and voila"), Ok((" and voila", "123")));
 /// assert_eq!(hex("DEADBEEF and others"), Ok((" and others", "DEADBEEF")));
 /// assert_eq!(hex("BADBABEsomething"), Ok(("something", "BADBABE")));
-/// assert_eq!(hex("D15EA5E"), Err(Err::Incomplete(Needed::new(1))));
-/// assert_eq!(hex(""), Err(Err::Incomplete(Needed::new(1))));
+/// assert_eq!(hex("D15EA5E"), Ok(("", "D15EA5E")));
+/// assert_eq!(hex(""), Err(Err::Error(Error::new("", ErrorKind::IsA))));
 /// ```
 pub fn is_a<T, I, Error: ParseError<I>>(arr: T) -> impl FnMut(I) -> IResult<I, I, Error>
 where
@@ -138,20 +143,18 @@ where
 {
   let mut parser = super::is_a(arr);
 
-  move |i: I| parser.process::<OutputM<Emit, Emit, Streaming>>(i)
+  move |i: I| parser.process::<OutputM<Emit, Emit, Complete>>(i)
 }
 
 /// Returns the longest input slice (if any) that matches the predicate.
 ///
 /// The parser will return the longest slice that matches the given predicate *(a function that
 /// takes the input and returns a bool)*.
-///
-/// # Streaming Specific
-/// *Streaming version* will return a `Err::Incomplete(Needed::new(1))` if the pattern reaches the end of the input.
 /// # Example
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, IResult};
-/// use nom::bytes::streaming::take_while;
+/// use nom::bytes::complete::take_while;
 /// use nom::AsChar;
 ///
 /// fn alpha(s: &[u8]) -> IResult<&[u8], &[u8]> {
@@ -160,8 +163,8 @@ where
 ///
 /// assert_eq!(alpha(b"latin123"), Ok((&b"123"[..], &b"latin"[..])));
 /// assert_eq!(alpha(b"12345"), Ok((&b"12345"[..], &b""[..])));
-/// assert_eq!(alpha(b"latin"), Err(Err::Incomplete(Needed::new(1))));
-/// assert_eq!(alpha(b""), Err(Err::Incomplete(Needed::new(1))));
+/// assert_eq!(alpha(b"latin"), Ok((&b""[..], &b"latin"[..])));
+/// assert_eq!(alpha(b""), Ok((&b""[..], &b""[..])));
 /// ```
 pub fn take_while<F, I, Error: ParseError<I>>(cond: F) -> impl FnMut(I) -> IResult<I, I, Error>
 where
@@ -170,7 +173,7 @@ where
 {
   let mut parser = super::take_while(cond);
 
-  move |i: I| parser.process::<OutputM<Emit, Emit, Streaming>>(i)
+  move |i: I| parser.process::<OutputM<Emit, Emit, Complete>>(i)
 }
 
 /// Returns the longest (at least 1) input slice that matches the predicate.
@@ -179,14 +182,11 @@ where
 /// takes the input and returns a bool)*.
 ///
 /// It will return an `Err(Err::Error((_, ErrorKind::TakeWhile1)))` if the pattern wasn't met.
-///
-/// # Streaming Specific
-/// *Streaming version* will return a `Err::Incomplete(Needed::new(1))` or if the pattern reaches the end of the input.
-///
 /// # Example
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::{Error, ErrorKind}, Needed, IResult};
-/// use nom::bytes::streaming::take_while1;
+/// use nom::bytes::complete::take_while1;
 /// use nom::AsChar;
 ///
 /// fn alpha(s: &[u8]) -> IResult<&[u8], &[u8]> {
@@ -194,7 +194,7 @@ where
 /// }
 ///
 /// assert_eq!(alpha(b"latin123"), Ok((&b"123"[..], &b"latin"[..])));
-/// assert_eq!(alpha(b"latin"), Err(Err::Incomplete(Needed::new(1))));
+/// assert_eq!(alpha(b"latin"), Ok((&b""[..], &b"latin"[..])));
 /// assert_eq!(alpha(b"12345"), Err(Err::Error(Error::new(&b"12345"[..], ErrorKind::TakeWhile1))));
 /// ```
 pub fn take_while1<F, I, Error: ParseError<I>>(cond: F) -> impl FnMut(I) -> IResult<I, I, Error>
@@ -204,22 +204,21 @@ where
 {
   let mut parser = super::take_while1(cond);
 
-  move |i: I| parser.process::<OutputM<Emit, Emit, Streaming>>(i)
+  move |i: I| parser.process::<OutputM<Emit, Emit, Complete>>(i)
 }
 
-/// Returns the longest (m <= len <= n) input slice  that matches the predicate.
+/// Returns the longest (m <= len <= n) input slice that matches the predicate.
 ///
 /// The parser will return the longest slice that matches the given predicate *(a function that
 /// takes the input and returns a bool)*.
 ///
-/// It will return an `Err::Error((_, ErrorKind::TakeWhileMN))` if the pattern wasn't met.
-/// # Streaming Specific
-/// *Streaming version* will return a `Err::Incomplete(Needed::new(1))`  if the pattern reaches the end of the input or is too short.
-///
+/// It will return an `Err::Error((_, ErrorKind::TakeWhileMN))` if the pattern wasn't met or is out
+/// of range (m <= len <= n).
 /// # Example
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::{Error, ErrorKind}, Needed, IResult};
-/// use nom::bytes::streaming::take_while_m_n;
+/// use nom::bytes::complete::take_while_m_n;
 /// use nom::AsChar;
 ///
 /// fn short_alpha(s: &[u8]) -> IResult<&[u8], &[u8]> {
@@ -228,8 +227,8 @@ where
 ///
 /// assert_eq!(short_alpha(b"latin123"), Ok((&b"123"[..], &b"latin"[..])));
 /// assert_eq!(short_alpha(b"lengthy"), Ok((&b"y"[..], &b"length"[..])));
-/// assert_eq!(short_alpha(b"latin"), Err(Err::Incomplete(Needed::new(1))));
-/// assert_eq!(short_alpha(b"ed"), Err(Err::Incomplete(Needed::new(1))));
+/// assert_eq!(short_alpha(b"latin"), Ok((&b""[..], &b"latin"[..])));
+/// assert_eq!(short_alpha(b"ed"), Err(Err::Error(Error::new(&b"ed"[..], ErrorKind::TakeWhileMN))));
 /// assert_eq!(short_alpha(b"12345"), Err(Err::Error(Error::new(&b"12345"[..], ErrorKind::TakeWhileMN))));
 /// ```
 pub fn take_while_m_n<F, I, Error: ParseError<I>>(
@@ -243,22 +242,18 @@ where
 {
   let mut parser = super::take_while_m_n(m, n, cond);
 
-  move |i: I| parser.process::<OutputM<Emit, Emit, Streaming>>(i)
+  move |i: I| parser.process::<OutputM<Emit, Emit, Complete>>(i)
 }
 
 /// Returns the longest input slice (if any) till a predicate is met.
 ///
 /// The parser will return the longest slice till the given predicate *(a function that
 /// takes the input and returns a bool)*.
-///
-/// # Streaming Specific
-/// *Streaming version* will return a `Err::Incomplete(Needed::new(1))` if the match reaches the
-/// end of input or if there was not match.
-///
 /// # Example
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, IResult};
-/// use nom::bytes::streaming::take_till;
+/// use nom::bytes::complete::take_till;
 ///
 /// fn till_colon(s: &str) -> IResult<&str, &str> {
 ///   take_till(|c| c == ':')(s)
@@ -266,8 +261,8 @@ where
 ///
 /// assert_eq!(till_colon("latin:123"), Ok((":123", "latin")));
 /// assert_eq!(till_colon(":empty matched"), Ok((":empty matched", ""))); //allowed
-/// assert_eq!(till_colon("12345"), Err(Err::Incomplete(Needed::new(1))));
-/// assert_eq!(till_colon(""), Err(Err::Incomplete(Needed::new(1))));
+/// assert_eq!(till_colon("12345"), Ok(("", "12345")));
+/// assert_eq!(till_colon(""), Ok(("", "")));
 /// ```
 #[allow(clippy::redundant_closure)]
 pub fn take_till<F, I, Error: ParseError<I>>(cond: F) -> impl FnMut(I) -> IResult<I, I, Error>
@@ -277,7 +272,7 @@ where
 {
   let mut parser = super::take_till(cond);
 
-  move |i: I| parser.process::<OutputM<Emit, Emit, Streaming>>(i)
+  move |i: I| parser.process::<OutputM<Emit, Emit, Complete>>(i)
 }
 
 /// Returns the longest (at least 1) input slice till a predicate is met.
@@ -285,13 +280,13 @@ where
 /// The parser will return the longest slice till the given predicate *(a function that
 /// takes the input and returns a bool)*.
 ///
-/// # Streaming Specific
-/// *Streaming version* will return a `Err::Incomplete(Needed::new(1))` if the match reaches the
-/// end of input or if there was not match.
+/// It will return `Err(Err::Error((_, ErrorKind::TakeTill1)))` if the input is empty or the
+/// predicate matches the first input.
 /// # Example
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::{Error, ErrorKind}, Needed, IResult};
-/// use nom::bytes::streaming::take_till1;
+/// use nom::bytes::complete::take_till1;
 ///
 /// fn till_colon(s: &str) -> IResult<&str, &str> {
 ///   take_till1(|c| c == ':')(s)
@@ -299,8 +294,8 @@ where
 ///
 /// assert_eq!(till_colon("latin:123"), Ok((":123", "latin")));
 /// assert_eq!(till_colon(":empty matched"), Err(Err::Error(Error::new(":empty matched", ErrorKind::TakeTill1))));
-/// assert_eq!(till_colon("12345"), Err(Err::Incomplete(Needed::new(1))));
-/// assert_eq!(till_colon(""), Err(Err::Incomplete(Needed::new(1))));
+/// assert_eq!(till_colon("12345"), Ok(("", "12345")));
+/// assert_eq!(till_colon(""), Err(Err::Error(Error::new("", ErrorKind::TakeTill1))));
 /// ```
 #[allow(clippy::redundant_closure)]
 pub fn take_till1<F, I, Error: ParseError<I>>(cond: F) -> impl FnMut(I) -> IResult<I, I, Error>
@@ -310,23 +305,17 @@ where
 {
   let mut parser = super::take_till1(cond);
 
-  move |i: I| parser.process::<OutputM<Emit, Emit, Streaming>>(i)
+  move |i: I| parser.process::<OutputM<Emit, Emit, Complete>>(i)
 }
 
 /// Returns an input slice containing the first N input elements (Input[..N]).
 ///
-/// # Streaming Specific
-/// *Streaming version* if the input has less than N elements, `take` will
-/// return a `Err::Incomplete(Needed::new(M))` where M is the number of
-/// additional bytes the parser would need to succeed.
-/// It is well defined for `&[u8]` as the number of elements is the byte size,
-/// but for types like `&str`, we cannot know how many bytes correspond for
-/// the next few chars, so the result will be `Err::Incomplete(Needed::Unknown)`
-///
+/// It will return `Err(Err::Error((_, ErrorKind::Eof)))` if the input is shorter than the argument.
 /// # Example
 /// ```rust
-/// # use nom::{Err, error::ErrorKind, Needed, IResult};
-/// use nom::bytes::streaming::take;
+/// # use nom_core as nom;
+/// # use nom::{Err, error::{Error, ErrorKind}, Needed, IResult};
+/// use nom::bytes::complete::take;
 ///
 /// fn take6(s: &str) -> IResult<&str, &str> {
 ///   take(6usize)(s)
@@ -334,7 +323,21 @@ where
 ///
 /// assert_eq!(take6("1234567"), Ok(("7", "123456")));
 /// assert_eq!(take6("things"), Ok(("", "things")));
-/// assert_eq!(take6("short"), Err(Err::Incomplete(Needed::Unknown)));
+/// assert_eq!(take6("short"), Err(Err::Error(Error::new("short", ErrorKind::Eof))));
+/// assert_eq!(take6(""), Err(Err::Error(Error::new("", ErrorKind::Eof))));
+/// ```
+///
+/// The units that are taken will depend on the input type. For example, for a
+/// `&str` it will take a number of `char`'s, whereas for a `&[u8]` it will
+/// take that many `u8`'s:
+///
+/// ```rust
+/// # use nom_core as nom;
+/// use nom::error::Error;
+/// use nom::bytes::complete::take;
+///
+/// assert_eq!(take::<_, _, Error<_>>(1usize)("💙"), Ok(("", "💙")));
+/// assert_eq!(take::<_, _, Error<_>>(1usize)("💙".as_bytes()), Ok((b"\x9F\x92\x99".as_ref(), b"\xF0".as_ref())));
 /// ```
 pub fn take<C, I, Error: ParseError<I>>(count: C) -> impl FnMut(I) -> IResult<I, I, Error>
 where
@@ -343,70 +346,66 @@ where
 {
   let mut parser = super::take(count);
 
-  move |i: I| parser.process::<OutputM<Emit, Emit, Streaming>>(i)
+  move |i: I| parser.process::<OutputM<Emit, Emit, Complete>>(i)
 }
 
 /// Returns the input slice up to the first occurrence of the pattern.
 ///
-/// It doesn't consume the pattern.
-///
-/// # Streaming Specific
-/// *Streaming version* will return a `Err::Incomplete(Needed::new(N))` if the input doesn't
-/// contain the pattern or if the input is smaller than the pattern.
+/// It doesn't consume the pattern. It will return `Err(Err::Error((_, ErrorKind::TakeUntil)))`
+/// if the pattern wasn't met.
 /// # Example
 /// ```rust
-/// # use nom::{Err, error::ErrorKind, Needed, IResult};
-/// use nom::bytes::streaming::take_until;
+/// # use nom_core as nom;
+/// # use nom::{Err, error::{Error, ErrorKind}, Needed, IResult};
+/// use nom::bytes::complete::take_until;
 ///
 /// fn until_eof(s: &str) -> IResult<&str, &str> {
 ///   take_until("eof")(s)
 /// }
 ///
 /// assert_eq!(until_eof("hello, worldeof"), Ok(("eof", "hello, world")));
-/// assert_eq!(until_eof("hello, world"), Err(Err::Incomplete(Needed::Unknown)));
-/// assert_eq!(until_eof("hello, worldeo"), Err(Err::Incomplete(Needed::Unknown)));
+/// assert_eq!(until_eof("hello, world"), Err(Err::Error(Error::new("hello, world", ErrorKind::TakeUntil))));
+/// assert_eq!(until_eof(""), Err(Err::Error(Error::new("", ErrorKind::TakeUntil))));
 /// assert_eq!(until_eof("1eof2eof"), Ok(("eof2eof", "1")));
 /// ```
 pub fn take_until<T, I, Error: ParseError<I>>(tag: T) -> impl FnMut(I) -> IResult<I, I, Error>
 where
   I: Input + FindSubstring<T>,
-  T: Clone,
+  T: Input + Clone,
 {
   let mut parser = super::take_until(tag);
 
-  move |i: I| parser.process::<OutputM<Emit, Emit, Streaming>>(i)
+  move |i: I| parser.process::<OutputM<Emit, Emit, Complete>>(i)
 }
 
 /// Returns the non empty input slice up to the first occurrence of the pattern.
 ///
-/// It doesn't consume the pattern.
-///
-/// # Streaming Specific
-/// *Streaming version* will return a `Err::Incomplete(Needed::new(N))` if the input doesn't
-/// contain the pattern or if the input is smaller than the pattern.
+/// It doesn't consume the pattern. It will return `Err(Err::Error((_, ErrorKind::TakeUntil)))`
+/// if the pattern wasn't met.
 /// # Example
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::{Error, ErrorKind}, Needed, IResult};
-/// use nom::bytes::streaming::take_until1;
+/// use nom::bytes::complete::take_until1;
 ///
 /// fn until_eof(s: &str) -> IResult<&str, &str> {
 ///   take_until1("eof")(s)
 /// }
 ///
 /// assert_eq!(until_eof("hello, worldeof"), Ok(("eof", "hello, world")));
-/// assert_eq!(until_eof("hello, world"), Err(Err::Incomplete(Needed::Unknown)));
-/// assert_eq!(until_eof("hello, worldeo"), Err(Err::Incomplete(Needed::Unknown)));
+/// assert_eq!(until_eof("hello, world"), Err(Err::Error(Error::new("hello, world", ErrorKind::TakeUntil))));
+/// assert_eq!(until_eof(""), Err(Err::Error(Error::new("", ErrorKind::TakeUntil))));
 /// assert_eq!(until_eof("1eof2eof"), Ok(("eof2eof", "1")));
-/// assert_eq!(until_eof("eof"),  Err(Err::Error(Error::new("eof", ErrorKind::TakeUntil))));
+/// assert_eq!(until_eof("eof"), Err(Err::Error(Error::new("eof", ErrorKind::TakeUntil))));
 /// ```
 pub fn take_until1<T, I, Error: ParseError<I>>(tag: T) -> impl FnMut(I) -> IResult<I, I, Error>
 where
   I: Input + FindSubstring<T>,
-  T: Clone,
+  T: Input + Clone,
 {
   let mut parser = super::take_until1(tag);
 
-  move |i: I| parser.process::<OutputM<Emit, Emit, Streaming>>(i)
+  move |i: I| parser.process::<OutputM<Emit, Emit, Complete>>(i)
 }
 
 /// Matches a byte string with escaped characters.
@@ -415,27 +414,28 @@ where
 /// * The second argument is the control character (like `\` in most languages)
 /// * The third argument matches the escaped characters
 /// # Example
-/// ```
+/// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, IResult};
 /// # use nom::character::complete::digit1;
-/// use nom::bytes::streaming::escaped;
-/// use nom::character::streaming::one_of;
+/// use nom::bytes::complete::escaped;
+/// use nom::character::complete::one_of;
 ///
 /// fn esc(s: &str) -> IResult<&str, &str> {
-///   escaped(digit1, '\\', one_of("\"n\\"))(s)
+///   escaped(digit1, '\\', one_of(r#""n\"#))(s)
 /// }
 ///
 /// assert_eq!(esc("123;"), Ok((";", "123")));
-/// assert_eq!(esc("12\\\"34;"), Ok((";", "12\\\"34")));
+/// assert_eq!(esc(r#"12\"34;"#), Ok((";", r#"12\"34"#)));
 /// ```
 ///
-pub fn escaped<I, Error, F, G>(
+pub fn escaped<'a, I: 'a, Error, F, G>(
   normal: F,
   control_char: char,
   escapable: G,
 ) -> impl FnMut(I) -> IResult<I, I, Error>
 where
-  I: Input + Clone + crate::traits::Offset,
+  I: Clone + crate::traits::Offset + Input,
   <I as Input>::Item: crate::traits::AsChar,
   F: Parser<I, Error = Error>,
   G: Parser<I, Error = Error>,
@@ -443,7 +443,7 @@ where
 {
   let mut parser = super::escaped(normal, control_char, escapable);
 
-  move |i: I| parser.process::<OutputM<Emit, Emit, Streaming>>(i)
+  move |i: I| parser.process::<OutputM<Emit, Emit, Complete>>(i)
 }
 
 /// Matches a byte string with escaped characters.
@@ -454,11 +454,12 @@ where
 ///
 /// As an example, the chain `abc\tdef` could be `abc    def` (it also consumes the control character)
 ///
-/// ```
+/// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, IResult};
 /// # use std::str::from_utf8;
-/// use nom::bytes::streaming::{escaped_transform, tag};
-/// use nom::character::streaming::alpha1;
+/// use nom::bytes::complete::{escaped_transform, tag};
+/// use nom::character::complete::alpha1;
 /// use nom::branch::alt;
 /// use nom::combinator::value;
 ///
@@ -474,7 +475,8 @@ where
 ///   )(input)
 /// }
 ///
-/// assert_eq!(parser("ab\\\"cd\""), Ok(("\"", String::from("ab\"cd"))));
+/// assert_eq!(parser("ab\\\"cd"), Ok(("", String::from("ab\"cd"))));
+/// assert_eq!(parser("ab\\ncd"), Ok(("", String::from("ab\ncd"))));
 /// ```
 #[cfg(feature = "alloc")]
 #[cfg_attr(feature = "docsrs", doc(cfg(feature = "alloc")))]
@@ -495,5 +497,85 @@ where
 {
   let mut parser = super::escaped_transform(normal, control_char, transform);
 
-  move |i: I| parser.process::<OutputM<Emit, Emit, Streaming>>(i)
+  move |i: I| parser.process::<OutputM<Emit, Emit, Complete>>(i)
+}
+
+#[cfg(test)]
+mod tests {
+  use super::*;
+  use crate::error::ErrorKind;
+  use crate::AsChar;
+  use crate::Err;
+
+  #[test]
+  fn complete_take_while_m_n_utf8_all_matching() {
+    let result: IResult<&str, &str> =
+      super::take_while_m_n(1, 4, |c: char| c.is_alphabetic())("øn");
+    assert_eq!(result, Ok(("", "øn")));
+  }
+
+  #[test]
+  fn complete_take_while_m_n_utf8_all_matching_substring() {
+    let result: IResult<&str, &str> =
+      super::take_while_m_n(1, 1, |c: char| c.is_alphabetic())("øn");
+    assert_eq!(result, Ok(("n", "ø")));
+  }
+
+  // issue #1336 "escaped hangs if normal parser accepts empty"
+  fn escaped_string(input: &str) -> IResult<&str, &str> {
+    use crate::character::complete::{alpha0, one_of};
+    escaped(alpha0, '\\', one_of("n"))(input)
+  }
+
+  // issue #1336 "escaped hangs if normal parser accepts empty"
+  #[test]
+  fn escaped_hang() {
+    escaped_string("7").unwrap();
+    escaped_string("a7").unwrap();
+  }
+
+  // issue ##1118 escaped does not work with empty string
+  fn unquote(input: &str) -> IResult<&str, &str> {
+    use crate::bytes::complete::*;
+    use crate::character::complete::*;
+    use crate::combinator::opt;
+    use crate::sequence::delimited;
+
+    delimited(
+      char('"'),
+      escaped(opt(none_of(r#"\""#)), '\\', one_of(r#"\"rnt"#)),
+      char('"'),
+    )
+    .parse(input)
+  }
+
+  #[test]
+  fn escaped_hang_1118() {
+    assert_eq!(unquote(r#""""#), Ok(("", "")));
+  }
+
+  // issue #1630 take_while_m_n is invalid for multi-byte UTF-8 characters
+  #[test]
+  fn complete_take_while_m_n_multibyte() {
+    use crate::error::Error;
+
+    fn multi_byte_chars(s: &str, m: usize, n: usize) -> IResult<&str, &str> {
+      take_while_m_n(m, n, |c: char| c.len() > 1)(s)
+    }
+
+    assert_eq!(multi_byte_chars("€ latin", 0, 64), Ok((" latin", "€")));
+    assert_eq!(multi_byte_chars("𝄠 latin", 0, 1), Ok((" latin", "𝄠")));
+    assert_eq!(multi_byte_chars("باب latin", 0, 64), Ok((" latin", "باب")));
+    assert_eq!(
+      multi_byte_chars("💣💢ᾠ latin", 3, 3),
+      Ok((" latin", "💣💢ᾠ"))
+    );
+    assert_eq!(multi_byte_chars("latin", 0, 64), Ok(("latin", "")));
+    assert_eq!(multi_byte_chars("باب", 1, 3), Ok(("", "باب")));
+    assert_eq!(multi_byte_chars("باب", 1, 2), Ok(("ب", "با")));
+    assert_eq!(
+      multi_byte_chars("latin", 1, 64),
+      Err(Err::Error(Error::new("latin", ErrorKind::TakeWhileMN)))
+    );
+  }
 }
