+++ nom_core/src/number/mod.rs
@@ -100,6 +100,7 @@ where
 /// Recognizes an unsigned 1 byte integer.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::be_u8;
 ///
@@ -121,6 +122,7 @@ where
 /// Recognizes a big endian unsigned 2 bytes integer.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::be_u16;
 ///
@@ -142,6 +144,7 @@ where
 /// Recognizes a big endian unsigned 3 byte integer.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::be_u24;
 ///
@@ -163,6 +166,7 @@ where
 /// Recognizes a big endian unsigned 4 bytes integer.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::be_u32;
 ///
@@ -184,6 +188,7 @@ where
 /// Recognizes a big endian unsigned 8 bytes integer.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::be_u64;
 ///
@@ -205,6 +210,7 @@ where
 /// Recognizes a big endian unsigned 16 bytes integer.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::be_u128;
 ///
@@ -227,6 +233,7 @@ where
 ///
 /// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::be_i8;
 ///
@@ -246,6 +253,7 @@ where
 /// Recognizes a big endian signed 2 bytes integer.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::be_i16;
 ///
@@ -265,6 +273,7 @@ where
 /// Recognizes a big endian signed 3 bytes integer.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::be_i24;
 ///
@@ -291,6 +300,7 @@ where
 /// Recognizes a big endian signed 4 bytes integer.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::be_i32;
 ///
@@ -310,6 +320,7 @@ where
 /// Recognizes a big endian signed 8 bytes integer.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::be_i64;
 ///
@@ -329,6 +340,7 @@ where
 /// Recognizes a big endian signed 16 bytes integer.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::be_i128;
 ///
@@ -408,6 +420,7 @@ where
 /// Recognizes an unsigned 1 byte integer.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::le_u8;
 ///
@@ -428,6 +441,7 @@ where
 ///
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::le_u16;
 ///
@@ -449,6 +463,7 @@ where
 /// Recognizes a little endian unsigned 3 bytes integer.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::le_u24;
 ///
@@ -470,6 +485,7 @@ where
 /// Recognizes a little endian unsigned 4 bytes integer.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::le_u32;
 ///
@@ -491,6 +507,7 @@ where
 /// Recognizes a little endian unsigned 8 bytes integer.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::le_u64;
 ///
@@ -512,6 +529,7 @@ where
 /// Recognizes a little endian unsigned 16 bytes integer.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::le_u128;
 ///
@@ -533,6 +551,7 @@ where
 /// Recognizes a signed 1 byte integer.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::le_i8;
 ///
@@ -552,6 +571,7 @@ where
 /// Recognizes a little endian signed 2 bytes integer.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::le_i16;
 ///
@@ -573,6 +593,7 @@ where
 /// Recognizes a little endian signed 3 bytes integer.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::le_i24;
 ///
@@ -601,6 +622,7 @@ where
 /// Recognizes a little endian signed 4 bytes integer.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::le_i32;
 ///
@@ -622,6 +644,7 @@ where
 /// Recognizes a little endian signed 8 bytes integer.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::le_i64;
 ///
@@ -643,6 +666,7 @@ where
 /// Recognizes a little endian signed 16 bytes integer.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::le_i128;
 ///
@@ -665,6 +689,7 @@ where
 ///
 /// Note that endianness does not apply to 1 byte numbers.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// # use nom::Needed::Size;
 /// use nom::number::u8;
@@ -690,6 +715,7 @@ where
 /// otherwise if `nom::number::Endianness::Little` parse a little endian u16 integer.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// # use nom::Needed::Size;
 /// use nom::number::u16;
@@ -730,6 +756,7 @@ where
 /// If the parameter is `nom::number::Endianness::Big`, parse a big endian u24 integer,
 /// otherwise if `nom::number::Endianness::Little` parse a little endian u24 integer.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// # use nom::Needed::Size;
 /// use nom::number::u24;
@@ -770,6 +797,7 @@ where
 /// If the parameter is `nom::number::Endianness::Big`, parse a big endian u32 integer,
 /// otherwise if `nom::number::Endianness::Little` parse a little endian u32 integer.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// # use nom::Needed::Size;
 /// use nom::number::u32;
@@ -810,6 +838,7 @@ where
 /// If the parameter is `nom::number::Endianness::Big`, parse a big endian u64 integer,
 /// otherwise if `nom::number::Endianness::Little` parse a little endian u64 integer.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// # use nom::Needed::Size;
 /// use nom::number::u64;
@@ -850,6 +879,7 @@ where
 /// If the parameter is `nom::number::Endianness::Big`, parse a big endian u128 integer,
 /// otherwise if `nom::number::Endianness::Little` parse a little endian u128 integer.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// # use nom::Needed::Size;
 /// use nom::number::u128;
@@ -890,6 +920,7 @@ where
 /// Note that endianness does not apply to 1 byte numbers.
 /// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// # use nom::Needed::Size;
 /// use nom::number::i8;
@@ -915,6 +946,7 @@ where
 /// otherwise if `nom::number::Endianness::Little` parse a little endian i16 integer.
 /// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// # use nom::Needed::Size;
 /// use nom::number::i16;
@@ -956,6 +988,7 @@ where
 /// otherwise if `nom::number::Endianness::Little` parse a little endian i24 integer.
 /// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// # use nom::Needed::Size;
 /// use nom::number::i24;
@@ -997,6 +1030,7 @@ where
 /// otherwise if `nom::number::Endianness::Little` parse a little endian i32 integer.
 /// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// # use nom::Needed::Size;
 /// use nom::number::i32;
@@ -1038,6 +1072,7 @@ where
 /// otherwise if `nom::number::Endianness::Little` parse a little endian i64 integer.
 /// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// # use nom::Needed::Size;
 /// use nom::number::i64;
@@ -1079,6 +1114,7 @@ where
 /// otherwise if `nom::number::Endianness::Little` parse a little endian i128 integer.
 /// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// # use nom::Needed::Size;
 /// use nom::number::i128;
@@ -1118,6 +1154,7 @@ where
 ///
 /// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::be_f32;
 ///
@@ -1140,6 +1177,7 @@ where
 ///
 /// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::be_f64;
 ///
@@ -1162,6 +1200,7 @@ where
 ///
 /// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::le_f32;
 ///
@@ -1184,6 +1223,7 @@ where
 ///
 /// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::le_f64;
 ///
@@ -1208,6 +1248,7 @@ where
 /// otherwise if `nom::number::Endianness::Little` parse a little endian f32 float.
 /// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// # use nom::Needed::Size;
 /// use nom::number::f32;
@@ -1249,6 +1290,7 @@ where
 /// otherwise if `nom::number::Endianness::Little` parse a little endian f64 float.
 /// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// # use nom::Needed::Size;
 /// use nom::number::f64;
@@ -1289,6 +1331,7 @@ where
 /// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if it reaches the end of input.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed, Parser};
 /// use nom::number::recognize_float;
 ///
