+++ nom_core/src/number/streaming.rs
@@ -1,30 +1,29 @@
-//! Parsers recognizing numbers, complete input version
+//! Parsers recognizing numbers, streaming version
 
 use crate::branch::alt;
-use crate::bytes::complete::tag;
-use crate::character::complete::{char, digit1, sign};
+use crate::bytes::streaming::tag;
+use crate::character::streaming::{char, digit1, sign};
 use crate::combinator::{cut, map, opt, recognize};
-use crate::error::ErrorKind;
-use crate::error::ParseError;
-use crate::internal::*;
+use crate::error::{ErrorKind, ParseError};
 use crate::lib::std::ops::{Add, Shl};
 use crate::sequence::pair;
-use crate::traits::{AsBytes, AsChar, Compare, Input, Offset};
+use crate::traits::{AsBytes, AsChar, Compare, Offset};
+use crate::{internal::*, Input};
 
 /// Recognizes an unsigned 1 byte integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::be_u8;
+/// use nom::number::streaming::be_u8;
 ///
 /// let parser = |s| {
-///   be_u8(s)
+///   be_u8::<_, (_, ErrorKind)>(s)
 /// };
 ///
-/// assert_eq!(parser(&b"\x00\x03abcefg"[..]), Ok((&b"\x03abcefg"[..], 0x00)));
-/// assert_eq!(parser(&b""[..]), Err(Err::Error((&[][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01abcd"[..]), Ok((&b"\x01abcd"[..], 0x00)));
+/// assert_eq!(parser(&b""[..]), Err(Err::Incomplete(Needed::new(1))));
 /// ```
 #[inline]
 pub fn be_u8<I, E: ParseError<I>>(input: I) -> IResult<I, u8, E>
@@ -36,18 +35,19 @@ where
 
 /// Recognizes a big endian unsigned 2 bytes integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
+///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::be_u16;
+/// use nom::number::streaming::be_u16;
 ///
 /// let parser = |s| {
-///   be_u16(s)
+///   be_u16::<_, (_, ErrorKind)>(s)
 /// };
 ///
-/// assert_eq!(parser(&b"\x00\x03abcefg"[..]), Ok((&b"abcefg"[..], 0x0003)));
-/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01abcd"[..]), Ok((&b"abcd"[..], 0x0001)));
+/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(1))));
 /// ```
 #[inline]
 pub fn be_u16<I, E: ParseError<I>>(input: I) -> IResult<I, u16, E>
@@ -59,18 +59,19 @@ where
 
 /// Recognizes a big endian unsigned 3 byte integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
+///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::be_u24;
+/// use nom::number::streaming::be_u24;
 ///
 /// let parser = |s| {
-///   be_u24(s)
+///   be_u24::<_, (_, ErrorKind)>(s)
 /// };
 ///
-/// assert_eq!(parser(&b"\x00\x03\x05abcefg"[..]), Ok((&b"abcefg"[..], 0x000305)));
-/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01\x02abcd"[..]), Ok((&b"abcd"[..], 0x000102)));
+/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(2))));
 /// ```
 #[inline]
 pub fn be_u24<I, E: ParseError<I>>(input: I) -> IResult<I, u32, E>
@@ -82,18 +83,19 @@ where
 
 /// Recognizes a big endian unsigned 4 bytes integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
+///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::be_u32;
+/// use nom::number::streaming::be_u32;
 ///
 /// let parser = |s| {
-///   be_u32(s)
+///   be_u32::<_, (_, ErrorKind)>(s)
 /// };
 ///
-/// assert_eq!(parser(&b"\x00\x03\x05\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x00030507)));
-/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01\x02\x03abcd"[..]), Ok((&b"abcd"[..], 0x00010203)));
+/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(3))));
 /// ```
 #[inline]
 pub fn be_u32<I, E: ParseError<I>>(input: I) -> IResult<I, u32, E>
@@ -105,18 +107,19 @@ where
 
 /// Recognizes a big endian unsigned 8 bytes integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
+///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::be_u64;
+/// use nom::number::streaming::be_u64;
 ///
 /// let parser = |s| {
-///   be_u64(s)
+///   be_u64::<_, (_, ErrorKind)>(s)
 /// };
 ///
-/// assert_eq!(parser(&b"\x00\x01\x02\x03\x04\x05\x06\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x0001020304050607)));
-/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01\x02\x03\x04\x05\x06\x07abcd"[..]), Ok((&b"abcd"[..], 0x0001020304050607)));
+/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(7))));
 /// ```
 #[inline]
 pub fn be_u64<I, E: ParseError<I>>(input: I) -> IResult<I, u64, E>
@@ -128,18 +131,18 @@ where
 
 /// Recognizes a big endian unsigned 16 bytes integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::be_u128;
+/// use nom::number::streaming::be_u128;
 ///
 /// let parser = |s| {
-///   be_u128(s)
+///   be_u128::<_, (_, ErrorKind)>(s)
 /// };
 ///
-/// assert_eq!(parser(&b"\x00\x01\x02\x03\x04\x05\x06\x07\x00\x01\x02\x03\x04\x05\x06\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x00010203040506070001020304050607)));
-/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01\x02\x03\x04\x05\x06\x07\x08\x09\x10\x11\x12\x13\x14\x15abcd"[..]), Ok((&b"abcd"[..], 0x00010203040506070809101112131415)));
+/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(15))));
 /// ```
 #[inline]
 pub fn be_u128<I, E: ParseError<I>>(input: I) -> IResult<I, u128, E>
@@ -155,23 +158,21 @@ where
   I: Input<Item = u8>,
   Uint: Default + Shl<u8, Output = Uint> + Add<Uint, Output = Uint> + From<u8>,
 {
-  super::be_uint(bound).parse_complete(input)
+  super::be_uint(bound).parse(input)
 }
 
 /// Recognizes a signed 1 byte integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::be_i8;
+/// use nom::number::streaming::be_i8;
 ///
-/// let parser = |s| {
-///   be_i8(s)
-/// };
+/// let parser = be_i8::<_, (_, ErrorKind)>;
 ///
-/// assert_eq!(parser(&b"\x00\x03abcefg"[..]), Ok((&b"\x03abcefg"[..], 0x00)));
-/// assert_eq!(parser(&b""[..]), Err(Err::Error((&[][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01abcd"[..]), Ok((&b"\x01abcd"[..], 0x00)));
+/// assert_eq!(parser(&b""[..]), Err(Err::Incomplete(Needed::new(1))));
 /// ```
 #[inline]
 pub fn be_i8<I, E: ParseError<I>>(input: I) -> IResult<I, i8, E>
@@ -183,18 +184,16 @@ where
 
 /// Recognizes a big endian signed 2 bytes integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::be_i16;
+/// use nom::number::streaming::be_i16;
 ///
-/// let parser = |s| {
-///   be_i16(s)
-/// };
+/// let parser = be_i16::<_, (_, ErrorKind)>;
 ///
-/// assert_eq!(parser(&b"\x00\x03abcefg"[..]), Ok((&b"abcefg"[..], 0x0003)));
-/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01abcd"[..]), Ok((&b"abcd"[..], 0x0001)));
+/// assert_eq!(parser(&b""[..]), Err(Err::Incomplete(Needed::new(2))));
 /// ```
 #[inline]
 pub fn be_i16<I, E: ParseError<I>>(input: I) -> IResult<I, i16, E>
@@ -206,18 +205,16 @@ where
 
 /// Recognizes a big endian signed 3 bytes integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::be_i24;
+/// use nom::number::streaming::be_i24;
 ///
-/// let parser = |s| {
-///   be_i24(s)
-/// };
+/// let parser = be_i24::<_, (_, ErrorKind)>;
 ///
-/// assert_eq!(parser(&b"\x00\x03\x05abcefg"[..]), Ok((&b"abcefg"[..], 0x000305)));
-/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01\x02abcd"[..]), Ok((&b"abcd"[..], 0x000102)));
+/// assert_eq!(parser(&b""[..]), Err(Err::Incomplete(Needed::new(3))));
 /// ```
 #[inline]
 pub fn be_i24<I, E: ParseError<I>>(input: I) -> IResult<I, i32, E>
@@ -238,18 +235,16 @@ where
 
 /// Recognizes a big endian signed 4 bytes integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::be_i32;
+/// use nom::number::streaming::be_i32;
 ///
-/// let parser = |s| {
-///   be_i32(s)
-/// };
+/// let parser = be_i32::<_, (_, ErrorKind)>;
 ///
-/// assert_eq!(parser(&b"\x00\x03\x05\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x00030507)));
-/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01\x02\x03abcd"[..]), Ok((&b"abcd"[..], 0x00010203)));
+/// assert_eq!(parser(&b""[..]), Err(Err::Incomplete(Needed::new(4))));
 /// ```
 #[inline]
 pub fn be_i32<I, E: ParseError<I>>(input: I) -> IResult<I, i32, E>
@@ -261,18 +256,17 @@ where
 
 /// Recognizes a big endian signed 8 bytes integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
+///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::be_i64;
+/// use nom::number::streaming::be_i64;
 ///
-/// let parser = |s| {
-///   be_i64(s)
-/// };
+/// let parser = be_i64::<_, (_, ErrorKind)>;
 ///
-/// assert_eq!(parser(&b"\x00\x01\x02\x03\x04\x05\x06\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x0001020304050607)));
-/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01\x02\x03\x04\x05\x06\x07abcd"[..]), Ok((&b"abcd"[..], 0x0001020304050607)));
+/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(7))));
 /// ```
 #[inline]
 pub fn be_i64<I, E: ParseError<I>>(input: I) -> IResult<I, i64, E>
@@ -284,18 +278,16 @@ where
 
 /// Recognizes a big endian signed 16 bytes integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::be_i128;
+/// use nom::number::streaming::be_i128;
 ///
-/// let parser = |s| {
-///   be_i128(s)
-/// };
+/// let parser = be_i128::<_, (_, ErrorKind)>;
 ///
-/// assert_eq!(parser(&b"\x00\x01\x02\x03\x04\x05\x06\x07\x00\x01\x02\x03\x04\x05\x06\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x00010203040506070001020304050607)));
-/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01\x02\x03\x04\x05\x06\x07\x08\x09\x10\x11\x12\x13\x14\x15abcd"[..]), Ok((&b"abcd"[..], 0x00010203040506070809101112131415)));
+/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(15))));
 /// ```
 #[inline]
 pub fn be_i128<I, E: ParseError<I>>(input: I) -> IResult<I, i128, E>
@@ -307,18 +299,16 @@ where
 
 /// Recognizes an unsigned 1 byte integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::le_u8;
+/// use nom::number::streaming::le_u8;
 ///
-/// let parser = |s| {
-///   le_u8(s)
-/// };
+/// let parser = le_u8::<_, (_, ErrorKind)>;
 ///
-/// assert_eq!(parser(&b"\x00\x03abcefg"[..]), Ok((&b"\x03abcefg"[..], 0x00)));
-/// assert_eq!(parser(&b""[..]), Err(Err::Error((&[][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01abcd"[..]), Ok((&b"\x01abcd"[..], 0x00)));
+/// assert_eq!(parser(&b""[..]), Err(Err::Incomplete(Needed::new(1))));
 /// ```
 #[inline]
 pub fn le_u8<I, E: ParseError<I>>(input: I) -> IResult<I, u8, E>
@@ -330,18 +320,19 @@ where
 
 /// Recognizes a little endian unsigned 2 bytes integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
+///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::le_u16;
+/// use nom::number::streaming::le_u16;
 ///
 /// let parser = |s| {
-///   le_u16(s)
+///   le_u16::<_, (_, ErrorKind)>(s)
 /// };
 ///
-/// assert_eq!(parser(&b"\x00\x03abcefg"[..]), Ok((&b"abcefg"[..], 0x0300)));
-/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01abcd"[..]), Ok((&b"abcd"[..], 0x0100)));
+/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(1))));
 /// ```
 #[inline]
 pub fn le_u16<I, E: ParseError<I>>(input: I) -> IResult<I, u16, E>
@@ -351,20 +342,21 @@ where
   le_uint(input, 2)
 }
 
-/// Recognizes a little endian unsigned 3 byte integer.
+/// Recognizes a little endian unsigned 3 bytes integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
+///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::le_u24;
+/// use nom::number::streaming::le_u24;
 ///
 /// let parser = |s| {
-///   le_u24(s)
+///   le_u24::<_, (_, ErrorKind)>(s)
 /// };
 ///
-/// assert_eq!(parser(&b"\x00\x03\x05abcefg"[..]), Ok((&b"abcefg"[..], 0x050300)));
-/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01\x02abcd"[..]), Ok((&b"abcd"[..], 0x020100)));
+/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(2))));
 /// ```
 #[inline]
 pub fn le_u24<I, E: ParseError<I>>(input: I) -> IResult<I, u32, E>
@@ -376,18 +368,19 @@ where
 
 /// Recognizes a little endian unsigned 4 bytes integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
+///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::le_u32;
+/// use nom::number::streaming::le_u32;
 ///
 /// let parser = |s| {
-///   le_u32(s)
+///   le_u32::<_, (_, ErrorKind)>(s)
 /// };
 ///
-/// assert_eq!(parser(&b"\x00\x03\x05\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x07050300)));
-/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01\x02\x03abcd"[..]), Ok((&b"abcd"[..], 0x03020100)));
+/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(3))));
 /// ```
 #[inline]
 pub fn le_u32<I, E: ParseError<I>>(input: I) -> IResult<I, u32, E>
@@ -399,18 +392,19 @@ where
 
 /// Recognizes a little endian unsigned 8 bytes integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
+///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::le_u64;
+/// use nom::number::streaming::le_u64;
 ///
 /// let parser = |s| {
-///   le_u64(s)
+///   le_u64::<_, (_, ErrorKind)>(s)
 /// };
 ///
-/// assert_eq!(parser(&b"\x00\x01\x02\x03\x04\x05\x06\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x0706050403020100)));
-/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01\x02\x03\x04\x05\x06\x07abcd"[..]), Ok((&b"abcd"[..], 0x0706050403020100)));
+/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(7))));
 /// ```
 #[inline]
 pub fn le_u64<I, E: ParseError<I>>(input: I) -> IResult<I, u64, E>
@@ -422,18 +416,19 @@ where
 
 /// Recognizes a little endian unsigned 16 bytes integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
+///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::le_u128;
+/// use nom::number::streaming::le_u128;
 ///
 /// let parser = |s| {
-///   le_u128(s)
+///   le_u128::<_, (_, ErrorKind)>(s)
 /// };
 ///
-/// assert_eq!(parser(&b"\x00\x01\x02\x03\x04\x05\x06\x07\x00\x01\x02\x03\x04\x05\x06\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x07060504030201000706050403020100)));
-/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01\x02\x03\x04\x05\x06\x07\x08\x09\x10\x11\x12\x13\x14\x15abcd"[..]), Ok((&b"abcd"[..], 0x15141312111009080706050403020100)));
+/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(15))));
 /// ```
 #[inline]
 pub fn le_u128<I, E: ParseError<I>>(input: I) -> IResult<I, u128, E>
@@ -449,46 +444,45 @@ where
   I: Input<Item = u8>,
   Uint: Default + Shl<u8, Output = Uint> + Add<Uint, Output = Uint> + From<u8>,
 {
-  super::le_uint(bound).parse_complete(input)
+  super::le_uint(bound).parse(input)
 }
 
 /// Recognizes a signed 1 byte integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::le_i8;
+/// use nom::number::streaming::le_i8;
 ///
-/// let parser = |s| {
-///   le_i8(s)
-/// };
+/// let parser = le_i8::<_, (_, ErrorKind)>;
 ///
-/// assert_eq!(parser(&b"\x00\x03abcefg"[..]), Ok((&b"\x03abcefg"[..], 0x00)));
-/// assert_eq!(parser(&b""[..]), Err(Err::Error((&[][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01abcd"[..]), Ok((&b"\x01abcd"[..], 0x00)));
+/// assert_eq!(parser(&b""[..]), Err(Err::Incomplete(Needed::new(1))));
 /// ```
 #[inline]
 pub fn le_i8<I, E: ParseError<I>>(input: I) -> IResult<I, i8, E>
 where
   I: Input<Item = u8>,
 {
-  be_u8.map(|x| x as i8).parse(input)
+  le_u8.map(|x| x as i8).parse(input)
 }
 
 /// Recognizes a little endian signed 2 bytes integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
+///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::le_i16;
+/// use nom::number::streaming::le_i16;
 ///
 /// let parser = |s| {
-///   le_i16(s)
+///   le_i16::<_, (_, ErrorKind)>(s)
 /// };
 ///
-/// assert_eq!(parser(&b"\x00\x03abcefg"[..]), Ok((&b"abcefg"[..], 0x0300)));
-/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01abcd"[..]), Ok((&b"abcd"[..], 0x0100)));
+/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(1))));
 /// ```
 #[inline]
 pub fn le_i16<I, E: ParseError<I>>(input: I) -> IResult<I, i16, E>
@@ -500,18 +494,19 @@ where
 
 /// Recognizes a little endian signed 3 bytes integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
+///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::le_i24;
+/// use nom::number::streaming::le_i24;
 ///
 /// let parser = |s| {
-///   le_i24(s)
+///   le_i24::<_, (_, ErrorKind)>(s)
 /// };
 ///
-/// assert_eq!(parser(&b"\x00\x03\x05abcefg"[..]), Ok((&b"abcefg"[..], 0x050300)));
-/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01\x02abcd"[..]), Ok((&b"abcd"[..], 0x020100)));
+/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(2))));
 /// ```
 #[inline]
 pub fn le_i24<I, E: ParseError<I>>(input: I) -> IResult<I, i32, E>
@@ -532,18 +527,19 @@ where
 
 /// Recognizes a little endian signed 4 bytes integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
+///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::le_i32;
+/// use nom::number::streaming::le_i32;
 ///
 /// let parser = |s| {
-///   le_i32(s)
+///   le_i32::<_, (_, ErrorKind)>(s)
 /// };
 ///
-/// assert_eq!(parser(&b"\x00\x03\x05\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x07050300)));
-/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01\x02\x03abcd"[..]), Ok((&b"abcd"[..], 0x03020100)));
+/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(3))));
 /// ```
 #[inline]
 pub fn le_i32<I, E: ParseError<I>>(input: I) -> IResult<I, i32, E>
@@ -555,18 +551,19 @@ where
 
 /// Recognizes a little endian signed 8 bytes integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
+///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::le_i64;
+/// use nom::number::streaming::le_i64;
 ///
 /// let parser = |s| {
-///   le_i64(s)
+///   le_i64::<_, (_, ErrorKind)>(s)
 /// };
 ///
-/// assert_eq!(parser(&b"\x00\x01\x02\x03\x04\x05\x06\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x0706050403020100)));
-/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01\x02\x03\x04\x05\x06\x07abcd"[..]), Ok((&b"abcd"[..], 0x0706050403020100)));
+/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(7))));
 /// ```
 #[inline]
 pub fn le_i64<I, E: ParseError<I>>(input: I) -> IResult<I, i64, E>
@@ -578,18 +575,19 @@ where
 
 /// Recognizes a little endian signed 16 bytes integer.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
+///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::le_i128;
+/// use nom::number::streaming::le_i128;
 ///
 /// let parser = |s| {
-///   le_i128(s)
+///   le_i128::<_, (_, ErrorKind)>(s)
 /// };
 ///
-/// assert_eq!(parser(&b"\x00\x01\x02\x03\x04\x05\x06\x07\x00\x01\x02\x03\x04\x05\x06\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x07060504030201000706050403020100)));
-/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b"\x00\x01\x02\x03\x04\x05\x06\x07\x08\x09\x10\x11\x12\x13\x14\x15abcd"[..]), Ok((&b"abcd"[..], 0x15141312111009080706050403020100)));
+/// assert_eq!(parser(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(15))));
 /// ```
 #[inline]
 pub fn le_i128<I, E: ParseError<I>>(input: I) -> IResult<I, i128, E>
@@ -602,51 +600,53 @@ where
 /// Recognizes an unsigned 1 byte integer
 ///
 /// Note that endianness does not apply to 1 byte numbers.
-/// *complete version*: returns an error if there is not enough input data
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
 /// # use nom::Needed::Size;
-/// use nom::number::complete::u8;
+/// use nom::number::streaming::u8;
 ///
 /// let parser = |s| {
-///   u8(s)
+///   u8::<_, (_, ErrorKind)>(s)
 /// };
 ///
 /// assert_eq!(parser(&b"\x00\x03abcefg"[..]), Ok((&b"\x03abcefg"[..], 0x00)));
-/// assert_eq!(parser(&b""[..]), Err(Err::Error((&[][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b""[..]), Err(Err::Incomplete(Needed::new(1))));
 /// ```
 #[inline]
 pub fn u8<I, E: ParseError<I>>(input: I) -> IResult<I, u8, E>
 where
   I: Input<Item = u8>,
 {
-  super::u8().parse_complete(input)
+  super::u8().parse(input)
 }
 
 /// Recognizes an unsigned 2 bytes integer
 ///
 /// If the parameter is `nom::number::Endianness::Big`, parse a big endian u16 integer,
 /// otherwise if `nom::number::Endianness::Little` parse a little endian u16 integer.
-/// *complete version*: returns an error if there is not enough input data
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
 /// # use nom::Needed::Size;
-/// use nom::number::complete::u16;
+/// use nom::number::streaming::u16;
 ///
 /// let be_u16 = |s| {
-///   u16(nom::number::Endianness::Big)(s)
+///   u16::<_, (_, ErrorKind)>(nom::number::Endianness::Big)(s)
 /// };
 ///
 /// assert_eq!(be_u16(&b"\x00\x03abcefg"[..]), Ok((&b"abcefg"[..], 0x0003)));
-/// assert_eq!(be_u16(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(be_u16(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(1))));
 ///
 /// let le_u16 = |s| {
-///   u16(nom::number::Endianness::Little)(s)
+///   u16::<_, (_, ErrorKind)>(nom::number::Endianness::Little)(s)
 /// };
 ///
 /// assert_eq!(le_u16(&b"\x00\x03abcefg"[..]), Ok((&b"abcefg"[..], 0x0300)));
-/// assert_eq!(le_u16(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(le_u16(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(1))));
 /// ```
 #[inline]
 pub fn u16<I, E: ParseError<I>>(
@@ -655,32 +655,33 @@ pub fn u16<I, E: ParseError<I>>(
 where
   I: Input<Item = u8>,
 {
-  move |input| super::u16(endian).parse_complete(input)
+  move |input| super::u16(endian).parse(input)
 }
 
 /// Recognizes an unsigned 3 byte integer
 ///
 /// If the parameter is `nom::number::Endianness::Big`, parse a big endian u24 integer,
 /// otherwise if `nom::number::Endianness::Little` parse a little endian u24 integer.
-/// *complete version*: returns an error if there is not enough input data
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
 /// # use nom::Needed::Size;
-/// use nom::number::complete::u24;
+/// use nom::number::streaming::u24;
 ///
 /// let be_u24 = |s| {
-///   u24(nom::number::Endianness::Big)(s)
+///   u24::<_,(_, ErrorKind)>(nom::number::Endianness::Big)(s)
 /// };
 ///
 /// assert_eq!(be_u24(&b"\x00\x03\x05abcefg"[..]), Ok((&b"abcefg"[..], 0x000305)));
-/// assert_eq!(be_u24(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(be_u24(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(2))));
 ///
 /// let le_u24 = |s| {
-///   u24(nom::number::Endianness::Little)(s)
+///   u24::<_, (_, ErrorKind)>(nom::number::Endianness::Little)(s)
 /// };
 ///
 /// assert_eq!(le_u24(&b"\x00\x03\x05abcefg"[..]), Ok((&b"abcefg"[..], 0x050300)));
-/// assert_eq!(le_u24(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(le_u24(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(2))));
 /// ```
 #[inline]
 pub fn u24<I, E: ParseError<I>>(
@@ -689,32 +690,33 @@ pub fn u24<I, E: ParseError<I>>(
 where
   I: Input<Item = u8>,
 {
-  move |input| super::u24(endian).parse_complete(input)
+  move |input| super::u24(endian).parse(input)
 }
 
 /// Recognizes an unsigned 4 byte integer
 ///
 /// If the parameter is `nom::number::Endianness::Big`, parse a big endian u32 integer,
 /// otherwise if `nom::number::Endianness::Little` parse a little endian u32 integer.
-/// *complete version*: returns an error if there is not enough input data
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
 /// # use nom::Needed::Size;
-/// use nom::number::complete::u32;
+/// use nom::number::streaming::u32;
 ///
 /// let be_u32 = |s| {
-///   u32(nom::number::Endianness::Big)(s)
+///   u32::<_, (_, ErrorKind)>(nom::number::Endianness::Big)(s)
 /// };
 ///
 /// assert_eq!(be_u32(&b"\x00\x03\x05\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x00030507)));
-/// assert_eq!(be_u32(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(be_u32(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(3))));
 ///
 /// let le_u32 = |s| {
-///   u32(nom::number::Endianness::Little)(s)
+///   u32::<_, (_, ErrorKind)>(nom::number::Endianness::Little)(s)
 /// };
 ///
 /// assert_eq!(le_u32(&b"\x00\x03\x05\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x07050300)));
-/// assert_eq!(le_u32(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(le_u32(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(3))));
 /// ```
 #[inline]
 pub fn u32<I, E: ParseError<I>>(
@@ -723,32 +725,33 @@ pub fn u32<I, E: ParseError<I>>(
 where
   I: Input<Item = u8>,
 {
-  move |input| super::u32(endian).parse_complete(input)
+  move |input| super::u32(endian).parse(input)
 }
 
 /// Recognizes an unsigned 8 byte integer
 ///
 /// If the parameter is `nom::number::Endianness::Big`, parse a big endian u64 integer,
 /// otherwise if `nom::number::Endianness::Little` parse a little endian u64 integer.
-/// *complete version*: returns an error if there is not enough input data
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
 /// # use nom::Needed::Size;
-/// use nom::number::complete::u64;
+/// use nom::number::streaming::u64;
 ///
 /// let be_u64 = |s| {
-///   u64(nom::number::Endianness::Big)(s)
+///   u64::<_, (_, ErrorKind)>(nom::number::Endianness::Big)(s)
 /// };
 ///
 /// assert_eq!(be_u64(&b"\x00\x01\x02\x03\x04\x05\x06\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x0001020304050607)));
-/// assert_eq!(be_u64(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(be_u64(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(7))));
 ///
 /// let le_u64 = |s| {
-///   u64(nom::number::Endianness::Little)(s)
+///   u64::<_, (_, ErrorKind)>(nom::number::Endianness::Little)(s)
 /// };
 ///
 /// assert_eq!(le_u64(&b"\x00\x01\x02\x03\x04\x05\x06\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x0706050403020100)));
-/// assert_eq!(le_u64(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(le_u64(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(7))));
 /// ```
 #[inline]
 pub fn u64<I, E: ParseError<I>>(
@@ -757,32 +760,33 @@ pub fn u64<I, E: ParseError<I>>(
 where
   I: Input<Item = u8>,
 {
-  move |input| super::u64(endian).parse_complete(input)
+  move |input| super::u64(endian).parse(input)
 }
 
 /// Recognizes an unsigned 16 byte integer
 ///
 /// If the parameter is `nom::number::Endianness::Big`, parse a big endian u128 integer,
 /// otherwise if `nom::number::Endianness::Little` parse a little endian u128 integer.
-/// *complete version*: returns an error if there is not enough input data
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
 /// # use nom::Needed::Size;
-/// use nom::number::complete::u128;
+/// use nom::number::streaming::u128;
 ///
 /// let be_u128 = |s| {
-///   u128(nom::number::Endianness::Big)(s)
+///   u128::<_, (_, ErrorKind)>(nom::number::Endianness::Big)(s)
 /// };
 ///
 /// assert_eq!(be_u128(&b"\x00\x01\x02\x03\x04\x05\x06\x07\x00\x01\x02\x03\x04\x05\x06\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x00010203040506070001020304050607)));
-/// assert_eq!(be_u128(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(be_u128(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(15))));
 ///
 /// let le_u128 = |s| {
-///   u128(nom::number::Endianness::Little)(s)
+///   u128::<_, (_, ErrorKind)>(nom::number::Endianness::Little)(s)
 /// };
 ///
 /// assert_eq!(le_u128(&b"\x00\x01\x02\x03\x04\x05\x06\x07\x00\x01\x02\x03\x04\x05\x06\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x07060504030201000706050403020100)));
-/// assert_eq!(le_u128(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(le_u128(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(15))));
 /// ```
 #[inline]
 pub fn u128<I, E: ParseError<I>>(
@@ -791,56 +795,58 @@ pub fn u128<I, E: ParseError<I>>(
 where
   I: Input<Item = u8>,
 {
-  move |input| super::u128(endian).parse_complete(input)
+  move |input| super::u128(endian).parse(input)
 }
 
 /// Recognizes a signed 1 byte integer
 ///
 /// Note that endianness does not apply to 1 byte numbers.
-/// *complete version*: returns an error if there is not enough input data
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
 /// # use nom::Needed::Size;
-/// use nom::number::complete::i8;
+/// use nom::number::streaming::i8;
 ///
 /// let parser = |s| {
-///   i8(s)
+///   i8::<_, (_, ErrorKind)>(s)
 /// };
 ///
 /// assert_eq!(parser(&b"\x00\x03abcefg"[..]), Ok((&b"\x03abcefg"[..], 0x00)));
-/// assert_eq!(parser(&b""[..]), Err(Err::Error((&[][..], ErrorKind::Eof))));
+/// assert_eq!(parser(&b""[..]), Err(Err::Incomplete(Needed::new(1))));
 /// ```
 #[inline]
 pub fn i8<I, E: ParseError<I>>(i: I) -> IResult<I, i8, E>
 where
   I: Input<Item = u8>,
 {
-  super::u8().map(|x| x as i8).parse_complete(i)
+  super::u8().map(|x| x as i8).parse(i)
 }
 
 /// Recognizes a signed 2 byte integer
 ///
 /// If the parameter is `nom::number::Endianness::Big`, parse a big endian i16 integer,
 /// otherwise if `nom::number::Endianness::Little` parse a little endian i16 integer.
-/// *complete version*: returns an error if there is not enough input data
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
 /// # use nom::Needed::Size;
-/// use nom::number::complete::i16;
+/// use nom::number::streaming::i16;
 ///
 /// let be_i16 = |s| {
-///   i16(nom::number::Endianness::Big)(s)
+///   i16::<_, (_, ErrorKind)>(nom::number::Endianness::Big)(s)
 /// };
 ///
 /// assert_eq!(be_i16(&b"\x00\x03abcefg"[..]), Ok((&b"abcefg"[..], 0x0003)));
-/// assert_eq!(be_i16(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(be_i16(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(1))));
 ///
 /// let le_i16 = |s| {
-///   i16(nom::number::Endianness::Little)(s)
+///   i16::<_, (_, ErrorKind)>(nom::number::Endianness::Little)(s)
 /// };
 ///
 /// assert_eq!(le_i16(&b"\x00\x03abcefg"[..]), Ok((&b"abcefg"[..], 0x0300)));
-/// assert_eq!(le_i16(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(le_i16(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(1))));
 /// ```
 #[inline]
 pub fn i16<I, E: ParseError<I>>(
@@ -849,32 +855,33 @@ pub fn i16<I, E: ParseError<I>>(
 where
   I: Input<Item = u8>,
 {
-  move |input| super::i16(endian).parse_complete(input)
+  move |input| super::i16(endian).parse(input)
 }
 
 /// Recognizes a signed 3 byte integer
 ///
 /// If the parameter is `nom::number::Endianness::Big`, parse a big endian i24 integer,
 /// otherwise if `nom::number::Endianness::Little` parse a little endian i24 integer.
-/// *complete version*: returns an error if there is not enough input data
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
 /// # use nom::Needed::Size;
-/// use nom::number::complete::i24;
+/// use nom::number::streaming::i24;
 ///
 /// let be_i24 = |s| {
-///   i24(nom::number::Endianness::Big)(s)
+///   i24::<_, (_, ErrorKind)>(nom::number::Endianness::Big)(s)
 /// };
 ///
 /// assert_eq!(be_i24(&b"\x00\x03\x05abcefg"[..]), Ok((&b"abcefg"[..], 0x000305)));
-/// assert_eq!(be_i24(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(be_i24(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(2))));
 ///
 /// let le_i24 = |s| {
-///   i24(nom::number::Endianness::Little)(s)
+///   i24::<_, (_, ErrorKind)>(nom::number::Endianness::Little)(s)
 /// };
 ///
 /// assert_eq!(le_i24(&b"\x00\x03\x05abcefg"[..]), Ok((&b"abcefg"[..], 0x050300)));
-/// assert_eq!(le_i24(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(le_i24(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(2))));
 /// ```
 #[inline]
 pub fn i24<I, E: ParseError<I>>(
@@ -883,32 +890,33 @@ pub fn i24<I, E: ParseError<I>>(
 where
   I: Input<Item = u8>,
 {
-  move |input| super::i24(endian).parse_complete(input)
+  move |input| super::i24(endian).parse(input)
 }
 
 /// Recognizes a signed 4 byte integer
 ///
 /// If the parameter is `nom::number::Endianness::Big`, parse a big endian i32 integer,
 /// otherwise if `nom::number::Endianness::Little` parse a little endian i32 integer.
-/// *complete version*: returns an error if there is not enough input data
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
 /// # use nom::Needed::Size;
-/// use nom::number::complete::i32;
+/// use nom::number::streaming::i32;
 ///
 /// let be_i32 = |s| {
-///   i32(nom::number::Endianness::Big)(s)
+///   i32::<_, (_, ErrorKind)>(nom::number::Endianness::Big)(s)
 /// };
 ///
 /// assert_eq!(be_i32(&b"\x00\x03\x05\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x00030507)));
-/// assert_eq!(be_i32(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(be_i32(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(3))));
 ///
 /// let le_i32 = |s| {
-///   i32(nom::number::Endianness::Little)(s)
+///   i32::<_, (_, ErrorKind)>(nom::number::Endianness::Little)(s)
 /// };
 ///
 /// assert_eq!(le_i32(&b"\x00\x03\x05\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x07050300)));
-/// assert_eq!(le_i32(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(le_i32(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(3))));
 /// ```
 #[inline]
 pub fn i32<I, E: ParseError<I>>(
@@ -917,32 +925,33 @@ pub fn i32<I, E: ParseError<I>>(
 where
   I: Input<Item = u8>,
 {
-  move |input| super::i32(endian).parse_complete(input)
+  move |input| super::i32(endian).parse(input)
 }
 
 /// Recognizes a signed 8 byte integer
 ///
 /// If the parameter is `nom::number::Endianness::Big`, parse a big endian i64 integer,
 /// otherwise if `nom::number::Endianness::Little` parse a little endian i64 integer.
-/// *complete version*: returns an error if there is not enough input data
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
 /// # use nom::Needed::Size;
-/// use nom::number::complete::i64;
+/// use nom::number::streaming::i64;
 ///
 /// let be_i64 = |s| {
-///   i64(nom::number::Endianness::Big)(s)
+///   i64::<_, (_, ErrorKind)>(nom::number::Endianness::Big)(s)
 /// };
 ///
 /// assert_eq!(be_i64(&b"\x00\x01\x02\x03\x04\x05\x06\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x0001020304050607)));
-/// assert_eq!(be_i64(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(be_i64(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(7))));
 ///
 /// let le_i64 = |s| {
-///   i64(nom::number::Endianness::Little)(s)
+///   i64::<_, (_, ErrorKind)>(nom::number::Endianness::Little)(s)
 /// };
 ///
 /// assert_eq!(le_i64(&b"\x00\x01\x02\x03\x04\x05\x06\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x0706050403020100)));
-/// assert_eq!(le_i64(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(le_i64(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(7))));
 /// ```
 #[inline]
 pub fn i64<I, E: ParseError<I>>(
@@ -951,32 +960,33 @@ pub fn i64<I, E: ParseError<I>>(
 where
   I: Input<Item = u8>,
 {
-  move |input| super::i64(endian).parse_complete(input)
+  move |input| super::i64(endian).parse(input)
 }
 
 /// Recognizes a signed 16 byte integer
 ///
 /// If the parameter is `nom::number::Endianness::Big`, parse a big endian i128 integer,
 /// otherwise if `nom::number::Endianness::Little` parse a little endian i128 integer.
-/// *complete version*: returns an error if there is not enough input data
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
 /// # use nom::Needed::Size;
-/// use nom::number::complete::i128;
+/// use nom::number::streaming::i128;
 ///
 /// let be_i128 = |s| {
-///   i128(nom::number::Endianness::Big)(s)
+///   i128::<_, (_, ErrorKind)>(nom::number::Endianness::Big)(s)
 /// };
 ///
 /// assert_eq!(be_i128(&b"\x00\x01\x02\x03\x04\x05\x06\x07\x00\x01\x02\x03\x04\x05\x06\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x00010203040506070001020304050607)));
-/// assert_eq!(be_i128(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(be_i128(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(15))));
 ///
 /// let le_i128 = |s| {
-///   i128(nom::number::Endianness::Little)(s)
+///   i128::<_, (_, ErrorKind)>(nom::number::Endianness::Little)(s)
 /// };
 ///
 /// assert_eq!(le_i128(&b"\x00\x01\x02\x03\x04\x05\x06\x07\x00\x01\x02\x03\x04\x05\x06\x07abcefg"[..]), Ok((&b"abcefg"[..], 0x07060504030201000706050403020100)));
-/// assert_eq!(le_i128(&b"\x01"[..]), Err(Err::Error((&[0x01][..], ErrorKind::Eof))));
+/// assert_eq!(le_i128(&b"\x01"[..]), Err(Err::Incomplete(Needed::new(15))));
 /// ```
 #[inline]
 pub fn i128<I, E: ParseError<I>>(
@@ -985,23 +995,23 @@ pub fn i128<I, E: ParseError<I>>(
 where
   I: Input<Item = u8>,
 {
-  move |input| super::i128(endian).parse_complete(input)
+  move |input| super::i128(endian).parse(input)
 }
 
 /// Recognizes a big endian 4 bytes floating point number.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::be_f32;
+/// use nom::number::streaming::be_f32;
 ///
 /// let parser = |s| {
-///   be_f32(s)
+///   be_f32::<_, (_, ErrorKind)>(s)
 /// };
 ///
-/// assert_eq!(parser(&[0x41, 0x48, 0x00, 0x00][..]), Ok((&b""[..], 12.5)));
-/// assert_eq!(parser(&b"abc"[..]), Err(Err::Error((&b"abc"[..], ErrorKind::Eof))));
+/// assert_eq!(parser(&[0x40, 0x29, 0x00, 0x00][..]), Ok((&b""[..], 2.640625)));
+/// assert_eq!(parser(&[0x01][..]), Err(Err::Incomplete(Needed::new(3))));
 /// ```
 #[inline]
 pub fn be_f32<I, E: ParseError<I>>(input: I) -> IResult<I, f32, E>
@@ -1016,18 +1026,18 @@ where
 
 /// Recognizes a big endian 8 bytes floating point number.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::be_f64;
+/// use nom::number::streaming::be_f64;
 ///
 /// let parser = |s| {
-///   be_f64(s)
+///   be_f64::<_, (_, ErrorKind)>(s)
 /// };
 ///
 /// assert_eq!(parser(&[0x40, 0x29, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00][..]), Ok((&b""[..], 12.5)));
-/// assert_eq!(parser(&b"abc"[..]), Err(Err::Error((&b"abc"[..], ErrorKind::Eof))));
+/// assert_eq!(parser(&[0x01][..]), Err(Err::Incomplete(Needed::new(7))));
 /// ```
 #[inline]
 pub fn be_f64<I, E: ParseError<I>>(input: I) -> IResult<I, f64, E>
@@ -1042,18 +1052,18 @@ where
 
 /// Recognizes a little endian 4 bytes floating point number.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::le_f32;
+/// use nom::number::streaming::le_f32;
 ///
 /// let parser = |s| {
-///   le_f32(s)
+///   le_f32::<_, (_, ErrorKind)>(s)
 /// };
 ///
 /// assert_eq!(parser(&[0x00, 0x00, 0x48, 0x41][..]), Ok((&b""[..], 12.5)));
-/// assert_eq!(parser(&b"abc"[..]), Err(Err::Error((&b"abc"[..], ErrorKind::Eof))));
+/// assert_eq!(parser(&[0x01][..]), Err(Err::Incomplete(Needed::new(3))));
 /// ```
 #[inline]
 pub fn le_f32<I, E: ParseError<I>>(input: I) -> IResult<I, f32, E>
@@ -1068,18 +1078,18 @@ where
 
 /// Recognizes a little endian 8 bytes floating point number.
 ///
-/// *Complete version*: Returns an error if there is not enough input data.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::le_f64;
+/// use nom::number::streaming::le_f64;
 ///
 /// let parser = |s| {
-///   le_f64(s)
+///   le_f64::<_, (_, ErrorKind)>(s)
 /// };
 ///
-/// assert_eq!(parser(&[0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x29, 0x40][..]), Ok((&b""[..], 12.5)));
-/// assert_eq!(parser(&b"abc"[..]), Err(Err::Error((&b"abc"[..], ErrorKind::Eof))));
+/// assert_eq!(parser(&[0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x48, 0x41][..]), Ok((&b""[..], 3145728.0)));
+/// assert_eq!(parser(&[0x01][..]), Err(Err::Incomplete(Needed::new(7))));
 /// ```
 #[inline]
 pub fn le_f64<I, E: ParseError<I>>(input: I) -> IResult<I, f64, E>
@@ -1096,25 +1106,26 @@ where
 ///
 /// If the parameter is `nom::number::Endianness::Big`, parse a big endian f32 float,
 /// otherwise if `nom::number::Endianness::Little` parse a little endian f32 float.
-/// *complete version*: returns an error if there is not enough input data
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
 /// # use nom::Needed::Size;
-/// use nom::number::complete::f32;
+/// use nom::number::streaming::f32;
 ///
 /// let be_f32 = |s| {
-///   f32(nom::number::Endianness::Big)(s)
+///   f32::<_, (_, ErrorKind)>(nom::number::Endianness::Big)(s)
 /// };
 ///
 /// assert_eq!(be_f32(&[0x41, 0x48, 0x00, 0x00][..]), Ok((&b""[..], 12.5)));
-/// assert_eq!(be_f32(&b"abc"[..]), Err(Err::Error((&b"abc"[..], ErrorKind::Eof))));
+/// assert_eq!(be_f32(&b"abc"[..]), Err(Err::Incomplete(Needed::new(1))));
 ///
 /// let le_f32 = |s| {
-///   f32(nom::number::Endianness::Little)(s)
+///   f32::<_, (_, ErrorKind)>(nom::number::Endianness::Little)(s)
 /// };
 ///
 /// assert_eq!(le_f32(&[0x00, 0x00, 0x48, 0x41][..]), Ok((&b""[..], 12.5)));
-/// assert_eq!(le_f32(&b"abc"[..]), Err(Err::Error((&b"abc"[..], ErrorKind::Eof))));
+/// assert_eq!(le_f32(&b"abc"[..]), Err(Err::Incomplete(Needed::new(1))));
 /// ```
 #[inline]
 pub fn f32<I, E: ParseError<I>>(endian: crate::number::Endianness) -> fn(I) -> IResult<I, f32, E>
@@ -1135,25 +1146,26 @@ where
 ///
 /// If the parameter is `nom::number::Endianness::Big`, parse a big endian f64 float,
 /// otherwise if `nom::number::Endianness::Little` parse a little endian f64 float.
-/// *complete version*: returns an error if there is not enough input data
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
 /// # use nom::Needed::Size;
-/// use nom::number::complete::f64;
+/// use nom::number::streaming::f64;
 ///
 /// let be_f64 = |s| {
-///   f64(nom::number::Endianness::Big)(s)
+///   f64::<_, (_, ErrorKind)>(nom::number::Endianness::Big)(s)
 /// };
 ///
 /// assert_eq!(be_f64(&[0x40, 0x29, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00][..]), Ok((&b""[..], 12.5)));
-/// assert_eq!(be_f64(&b"abc"[..]), Err(Err::Error((&b"abc"[..], ErrorKind::Eof))));
+/// assert_eq!(be_f64(&b"abc"[..]), Err(Err::Incomplete(Needed::new(5))));
 ///
 /// let le_f64 = |s| {
-///   f64(nom::number::Endianness::Little)(s)
+///   f64::<_, (_, ErrorKind)>(nom::number::Endianness::Little)(s)
 /// };
 ///
 /// assert_eq!(le_f64(&[0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x29, 0x40][..]), Ok((&b""[..], 12.5)));
-/// assert_eq!(le_f64(&b"abc"[..]), Err(Err::Error((&b"abc"[..], ErrorKind::Eof))));
+/// assert_eq!(le_f64(&b"abc"[..]), Err(Err::Incomplete(Needed::new(5))));
 /// ```
 #[inline]
 pub fn f64<I, E: ParseError<I>>(endian: crate::number::Endianness) -> fn(I) -> IResult<I, f64, E>
@@ -1172,29 +1184,28 @@ where
 
 /// Recognizes a hex-encoded integer.
 ///
-/// *Complete version*: Will parse until the end of input if it has less than 8 bytes.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::hex_u32;
+/// use nom::number::streaming::hex_u32;
 ///
 /// let parser = |s| {
 ///   hex_u32(s)
 /// };
 ///
-/// assert_eq!(parser(&b"01AE"[..]), Ok((&b""[..], 0x01AE)));
-/// assert_eq!(parser(&b"abc"[..]), Ok((&b""[..], 0x0ABC)));
+/// assert_eq!(parser(&b"01AE;"[..]), Ok((&b";"[..], 0x01AE)));
+/// assert_eq!(parser(&b"abc"[..]), Err(Err::Incomplete(Needed::new(1))));
 /// assert_eq!(parser(&b"ggg"[..]), Err(Err::Error((&b"ggg"[..], ErrorKind::IsA))));
 /// ```
 #[inline]
 pub fn hex_u32<I, E: ParseError<I>>(input: I) -> IResult<I, u32, E>
 where
-  I: Input,
+  I: Input + AsBytes,
   <I as Input>::Item: AsChar,
-  I: AsBytes,
 {
   let e: ErrorKind = ErrorKind::IsA;
-  let (i, o) = input.split_at_position1_complete(
+  let (i, o) = input.split_at_position1(
     |c| {
       let c = c.as_char();
       !"0123456789abcdefABCDEF".contains(c)
@@ -1223,21 +1234,21 @@ where
   Ok((remaining, res))
 }
 
-/// Recognizes floating point number in a byte string and returns the corresponding slice.
+/// Recognizes a floating point number in text format and returns the corresponding part of the input.
 ///
-/// *Complete version*: Can parse until the end of input.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if it reaches the end of input.
 ///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
-/// # use nom::Needed::Size;
-/// use nom::number::complete::recognize_float;
+/// use nom::number::streaming::recognize_float;
 ///
 /// let parser = |s| {
 ///   recognize_float(s)
 /// };
 ///
-/// assert_eq!(parser("11e-1"), Ok(("", "11e-1")));
-/// assert_eq!(parser("123E-02"), Ok(("", "123E-02")));
+/// assert_eq!(parser("11e-1;"), Ok((";", "11e-1")));
+/// assert_eq!(parser("123E-02;"), Ok((";", "123E-02")));
 /// assert_eq!(parser("123K-01"), Ok(("K-01", "123")));
 /// assert_eq!(parser("abc"), Err(Err::Error(("abc", ErrorKind::Char))));
 /// ```
@@ -1249,7 +1260,7 @@ where
   <T as Input>::Item: AsChar,
 {
   recognize((
-    opt(alt((char('+'), char('-')))),
+      opt(alt((char('+'), char('-')))),
       alt((
         map((digit1, opt(pair(char('.'), opt(digit1)))), |_| ()),
         map((char('.'), digit1), |_| ())
@@ -1279,15 +1290,15 @@ where
       })
     },
     |i: T| {
-      crate::bytes::complete::tag_no_case::<_, _, E>("nan")(i.clone())
+      crate::bytes::streaming::tag_no_case::<_, _, E>("nan")(i.clone())
         .map_err(|_| crate::Err::Error(E::from_error_kind(i, ErrorKind::Float)))
     },
     |i: T| {
-      crate::bytes::complete::tag_no_case::<_, _, E>("infinity")(i.clone())
+      crate::bytes::streaming::tag_no_case::<_, _, E>("infinity")(i.clone())
         .map_err(|_| crate::Err::Error(E::from_error_kind(i, ErrorKind::Float)))
     },
     |i: T| {
-      crate::bytes::complete::tag_no_case::<_, _, E>("inf")(i.clone())
+      crate::bytes::streaming::tag_no_case::<_, _, E>("inf")(i.clone())
         .map_err(|_| crate::Err::Error(E::from_error_kind(i, ErrorKind::Float)))
     },
   ))
@@ -1299,7 +1310,7 @@ where
 /// It returns a tuple of (`sign`, `integer part`, `fraction part` and `exponent`) of the input
 /// data.
 ///
-/// *Complete version*: Can parse until the end of input.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
 ///
 pub fn recognize_float_parts<T, E: ParseError<T>>(input: T) -> IResult<T, (bool, T, T, i32), E>
 where
@@ -1311,11 +1322,12 @@ where
 {
   let (i, sign) = sign(input.clone())?;
 
-  //let (i, zeroes) = take_while(|c: <T as Input>::Item| c.as_char() == '0')(i)?;
+  //let (i, zeroes) = take_while(|c: <T as InputTakeAtPosition>::Item| c.as_char() == '0')(i)?;
   let (i, zeroes) = match i.as_bytes().iter().position(|c| *c != b'0') {
     Some(index) => i.take_split(index),
     None => i.take_split(i.input_len()),
   };
+
   //let (i, mut integer) = digit0(i)?;
   let (i, mut integer) = match i
     .as_bytes()
@@ -1352,8 +1364,10 @@ where
       }
     }
 
-    #[allow(clippy::or_fun_call)]
-    let position = position.unwrap_or(i.input_len());
+    let position = match position {
+      Some(p) => p,
+      None => return Err(Err::Incomplete(Needed::new(1))),
+    };
 
     let index = if zero_count == 0 {
       position
@@ -1378,7 +1392,7 @@ where
   };
 
   let (i, exp) = if e {
-    cut(crate::character::complete::i32).parse(i)?
+    cut(crate::character::streaming::i32).parse(i)?
   } else {
     (i2, 0)
   };
@@ -1386,12 +1400,12 @@ where
   Ok((i, (sign, integer, fraction, exp)))
 }
 
-use crate::traits::ParseTo;
-
 /// Recognizes floating point number in text format and returns a f32.
 ///
-/// *Complete version*: Can parse until the end of input.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
+///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
 /// # use nom::Needed::Size;
 /// use nom::number::complete::float;
@@ -1407,10 +1421,9 @@ use crate::traits::ParseTo;
 /// ```
 pub fn float<T, E: ParseError<T>>(input: T) -> IResult<T, f32, E>
 where
-  T: Clone + Offset + ParseTo<f32> + Compare<&'static str>,
-  T: Input,
-  <T as Input>::Item: AsChar,
-  <T as Input>::Iter: Clone,
+  T: Clone + Offset,
+  T: Input + crate::traits::ParseTo<f32> + Compare<&'static str>,
+  <T as Input>::Item: AsChar + Clone,
   T: AsBytes,
   T: for<'a> Compare<&'a [u8]>,
 {
@@ -1427,7 +1440,7 @@ where
   }
 
   Ok((i, float))
-      */
+  */
   let (i, s) = recognize_float_or_exceptions(input)?;
   match s.parse_to() {
     Some(f) => Ok((i, f)),
@@ -1440,8 +1453,10 @@ where
 
 /// Recognizes floating point number in text format and returns a f64.
 ///
-/// *Complete version*: Can parse until the end of input.
+/// *Streaming version*: Will return `Err(nom::Err::Incomplete(_))` if there is not enough data.
+///
 /// ```rust
+/// # use nom_core as nom;
 /// # use nom::{Err, error::ErrorKind, Needed};
 /// # use nom::Needed::Size;
 /// use nom::number::complete::double;
@@ -1457,10 +1472,9 @@ where
 /// ```
 pub fn double<T, E: ParseError<T>>(input: T) -> IResult<T, f64, E>
 where
-  T: Clone + Offset + ParseTo<f64> + Compare<&'static str>,
-  T: Input,
-  <T as Input>::Item: AsChar,
-  <T as Input>::Iter: Clone,
+  T: Clone + Offset,
+  T: Input + crate::traits::ParseTo<f64> + Compare<&'static str>,
+  <T as Input>::Item: AsChar + Clone,
   T: AsBytes,
   T: for<'a> Compare<&'a [u8]>,
 {
@@ -1477,7 +1491,7 @@ where
   }
 
   Ok((i, float))
-      */
+  */
   let (i, s) = recognize_float_or_exceptions(input)?;
   match s.parse_to() {
     Some(f) => Ok((i, f)),
@@ -1492,7 +1506,7 @@ where
 mod tests {
   use super::*;
   use crate::error::ErrorKind;
-  use crate::internal::Err;
+  use crate::internal::{Err, Needed};
   use proptest::prelude::*;
 
   macro_rules! assert_parse(
@@ -1504,50 +1518,57 @@ mod tests {
 
   #[test]
   fn i8_tests() {
-    assert_parse!(i8(&[0x00][..]), Ok((&b""[..], 0)));
-    assert_parse!(i8(&[0x7f][..]), Ok((&b""[..], 127)));
-    assert_parse!(i8(&[0xff][..]), Ok((&b""[..], -1)));
-    assert_parse!(i8(&[0x80][..]), Ok((&b""[..], -128)));
-  }
-
-  #[test]
-  fn be_i8_tests() {
     assert_parse!(be_i8(&[0x00][..]), Ok((&b""[..], 0)));
     assert_parse!(be_i8(&[0x7f][..]), Ok((&b""[..], 127)));
     assert_parse!(be_i8(&[0xff][..]), Ok((&b""[..], -1)));
     assert_parse!(be_i8(&[0x80][..]), Ok((&b""[..], -128)));
+    assert_parse!(be_i8(&[][..]), Err(Err::Incomplete(Needed::new(1))));
   }
 
   #[test]
-  fn be_i16_tests() {
+  fn i16_tests() {
     assert_parse!(be_i16(&[0x00, 0x00][..]), Ok((&b""[..], 0)));
     assert_parse!(be_i16(&[0x7f, 0xff][..]), Ok((&b""[..], 32_767_i16)));
     assert_parse!(be_i16(&[0xff, 0xff][..]), Ok((&b""[..], -1)));
     assert_parse!(be_i16(&[0x80, 0x00][..]), Ok((&b""[..], -32_768_i16)));
+    assert_parse!(be_i16(&[][..]), Err(Err::Incomplete(Needed::new(2))));
+    assert_parse!(be_i16(&[0x00][..]), Err(Err::Incomplete(Needed::new(1))));
   }
 
   #[test]
-  fn be_u24_tests() {
+  fn u24_tests() {
     assert_parse!(be_u24(&[0x00, 0x00, 0x00][..]), Ok((&b""[..], 0)));
     assert_parse!(be_u24(&[0x00, 0xFF, 0xFF][..]), Ok((&b""[..], 65_535_u32)));
     assert_parse!(
       be_u24(&[0x12, 0x34, 0x56][..]),
       Ok((&b""[..], 1_193_046_u32))
     );
+    assert_parse!(be_u24(&[][..]), Err(Err::Incomplete(Needed::new(3))));
+    assert_parse!(be_u24(&[0x00][..]), Err(Err::Incomplete(Needed::new(2))));
+    assert_parse!(
+      be_u24(&[0x00, 0x00][..]),
+      Err(Err::Incomplete(Needed::new(1)))
+    );
   }
 
   #[test]
-  fn be_i24_tests() {
+  fn i24_tests() {
     assert_parse!(be_i24(&[0xFF, 0xFF, 0xFF][..]), Ok((&b""[..], -1_i32)));
     assert_parse!(be_i24(&[0xFF, 0x00, 0x00][..]), Ok((&b""[..], -65_536_i32)));
     assert_parse!(
       be_i24(&[0xED, 0xCB, 0xAA][..]),
       Ok((&b""[..], -1_193_046_i32))
     );
+    assert_parse!(be_i24(&[][..]), Err(Err::Incomplete(Needed::new(3))));
+    assert_parse!(be_i24(&[0x00][..]), Err(Err::Incomplete(Needed::new(2))));
+    assert_parse!(
+      be_i24(&[0x00, 0x00][..]),
+      Err(Err::Incomplete(Needed::new(1)))
+    );
   }
 
   #[test]
-  fn be_i32_tests() {
+  fn i32_tests() {
     assert_parse!(be_i32(&[0x00, 0x00, 0x00, 0x00][..]), Ok((&b""[..], 0)));
     assert_parse!(
       be_i32(&[0x7f, 0xff, 0xff, 0xff][..]),
@@ -1558,10 +1579,20 @@ mod tests {
       be_i32(&[0x80, 0x00, 0x00, 0x00][..]),
       Ok((&b""[..], -2_147_483_648_i32))
     );
+    assert_parse!(be_i32(&[][..]), Err(Err::Incomplete(Needed::new(4))));
+    assert_parse!(be_i32(&[0x00][..]), Err(Err::Incomplete(Needed::new(3))));
+    assert_parse!(
+      be_i32(&[0x00, 0x00][..]),
+      Err(Err::Incomplete(Needed::new(2)))
+    );
+    assert_parse!(
+      be_i32(&[0x00, 0x00, 0x00][..]),
+      Err(Err::Incomplete(Needed::new(1)))
+    );
   }
 
   #[test]
-  fn be_i64_tests() {
+  fn i64_tests() {
     assert_parse!(
       be_i64(&[0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00][..]),
       Ok((&b""[..], 0))
@@ -1578,10 +1609,36 @@ mod tests {
       be_i64(&[0x80, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00][..]),
       Ok((&b""[..], -9_223_372_036_854_775_808_i64))
     );
+    assert_parse!(be_i64(&[][..]), Err(Err::Incomplete(Needed::new(8))));
+    assert_parse!(be_i64(&[0x00][..]), Err(Err::Incomplete(Needed::new(7))));
+    assert_parse!(
+      be_i64(&[0x00, 0x00][..]),
+      Err(Err::Incomplete(Needed::new(6)))
+    );
+    assert_parse!(
+      be_i64(&[0x00, 0x00, 0x00][..]),
+      Err(Err::Incomplete(Needed::new(5)))
+    );
+    assert_parse!(
+      be_i64(&[0x00, 0x00, 0x00, 0x00][..]),
+      Err(Err::Incomplete(Needed::new(4)))
+    );
+    assert_parse!(
+      be_i64(&[0x00, 0x00, 0x00, 0x00, 0x00][..]),
+      Err(Err::Incomplete(Needed::new(3)))
+    );
+    assert_parse!(
+      be_i64(&[0x00, 0x00, 0x00, 0x00, 0x00, 0x00][..]),
+      Err(Err::Incomplete(Needed::new(2)))
+    );
+    assert_parse!(
+      be_i64(&[0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00][..]),
+      Err(Err::Incomplete(Needed::new(1)))
+    );
   }
 
   #[test]
-  fn be_i128_tests() {
+  fn i128_tests() {
     assert_parse!(
       be_i128(
         &[
@@ -1623,6 +1680,69 @@ mod tests {
         &b""[..],
         -170_141_183_460_469_231_731_687_303_715_884_105_728_i128
       ))
+    );
+    assert_parse!(be_i128(&[][..]), Err(Err::Incomplete(Needed::new(16))));
+    assert_parse!(be_i128(&[0x00][..]), Err(Err::Incomplete(Needed::new(15))));
+    assert_parse!(
+      be_i128(&[0x00, 0x00][..]),
+      Err(Err::Incomplete(Needed::new(14)))
+    );
+    assert_parse!(
+      be_i128(&[0x00, 0x00, 0x00][..]),
+      Err(Err::Incomplete(Needed::new(13)))
+    );
+    assert_parse!(
+      be_i128(&[0x00, 0x00, 0x00, 0x00][..]),
+      Err(Err::Incomplete(Needed::new(12)))
+    );
+    assert_parse!(
+      be_i128(&[0x00, 0x00, 0x00, 0x00, 0x00][..]),
+      Err(Err::Incomplete(Needed::new(11)))
+    );
+    assert_parse!(
+      be_i128(&[0x00, 0x00, 0x00, 0x00, 0x00, 0x00][..]),
+      Err(Err::Incomplete(Needed::new(10)))
+    );
+    assert_parse!(
+      be_i128(&[0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00][..]),
+      Err(Err::Incomplete(Needed::new(9)))
+    );
+    assert_parse!(
+      be_i128(&[0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00][..]),
+      Err(Err::Incomplete(Needed::new(8)))
+    );
+    assert_parse!(
+      be_i128(&[0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00][..]),
+      Err(Err::Incomplete(Needed::new(7)))
+    );
+    assert_parse!(
+      be_i128(&[0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00][..]),
+      Err(Err::Incomplete(Needed::new(6)))
+    );
+    assert_parse!(
+      be_i128(&[0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00][..]),
+      Err(Err::Incomplete(Needed::new(5)))
+    );
+    assert_parse!(
+      be_i128(&[0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00][..]),
+      Err(Err::Incomplete(Needed::new(4)))
+    );
+    assert_parse!(
+      be_i128(&[0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00][..]),
+      Err(Err::Incomplete(Needed::new(3)))
+    );
+    assert_parse!(
+      be_i128(
+        &[0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00][..]
+      ),
+      Err(Err::Incomplete(Needed::new(2)))
+    );
+    assert_parse!(
+      be_i128(
+        &[0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00]
+          [..]
+      ),
+      Err(Err::Incomplete(Needed::new(1)))
     );
   }
 
@@ -1801,7 +1921,7 @@ mod tests {
     );
     assert_parse!(hex_u32(&b"ffffffff;"[..]), Ok((&b";"[..], 4_294_967_295)));
     assert_parse!(hex_u32(&b"0x1be2;"[..]), Ok((&b"x1be2;"[..], 0)));
-    assert_parse!(hex_u32(&b"12af"[..]), Ok((&b""[..], 0x12af)));
+    assert_parse!(hex_u32(&b"12af"[..]), Err(Err::Incomplete(Needed::new(1))));
   }
 
   #[test]
@@ -1833,19 +1953,20 @@ mod tests {
 
       println!("now parsing: {} -> {}", test, expected32);
 
-      assert_parse!(recognize_float(test), Ok(("", test)));
+      let larger = format!("{};", test);
+      assert_parse!(recognize_float(&larger[..]), Ok((";", test)));
 
-      assert_parse!(float(test.as_bytes()), Ok((&b""[..], expected32)));
-      assert_parse!(float(test), Ok(("", expected32)));
+      assert_parse!(float(larger.as_bytes()), Ok((&b";"[..], expected32)));
+      assert_parse!(float(&larger[..]), Ok((";", expected32)));
 
-      assert_parse!(double(test.as_bytes()), Ok((&b""[..], expected64)));
-      assert_parse!(double(test), Ok(("", expected64)));
+      assert_parse!(double(larger.as_bytes()), Ok((&b";"[..], expected64)));
+      assert_parse!(double(&larger[..]), Ok((";", expected64)));
     }
 
     let remaining_exponent = "-1.234E-";
     assert_parse!(
       recognize_float(remaining_exponent),
-      Err(Err::Failure(("", ErrorKind::Digit)))
+      Err(Err::Incomplete(Needed::new(1)))
     );
 
     let (_i, nan) = float::<_, ()>("NaN").unwrap();
@@ -1943,6 +2064,7 @@ mod tests {
 
   #[cfg(feature = "std")]
   fn parse_f64(i: &str) -> IResult<&str, f64, ()> {
+    use crate::traits::ParseTo;
     match recognize_float_or_exceptions(i) {
       Err(e) => Err(e),
       Ok((i, s)) => {
